- Add credits to Cephes - use fastabs for complex numbers - add a test for asymptotic behaviour from issue #4301
