+++ test/math.jl
@@ -459,6 +459,8 @@ end
       0.00634645247782269506319336871208405439180447035257028310080 -
       0.00169495384841964531409376316336552555952269360134349446910im)
 
+@test_approx_eq lbeta(1e10, 1e-10) 23.02585092758015 # comment in issue #4301
+
  for i=1:10
      @test beta(-i, i) == (-1)^i / i
  end
